Drop: aapt and aapt2's selective configuration
